requirements.txt
requests
bs4
pyserial
esptool

